requirements.txt
matplotlib==3.4.3
nltk==3.9.1
numpy==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.5
PyMuPDF==1.25.3
pyparsing==3.2.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
pywin32==308
PyYAML==6.0.2
pyzotero==1.6.8
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.22.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.21.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
wordninja==2.0.0
yarg==0.1.9
zipp==3.21.0
pyzmq==26.2.1
